+++ CHANGELOG.md
@@ -4,8 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `buttonText` prop to `CrowdfundingCard`.
-- Fix: Fix overflow style on `ScrollableContainer`.
+## [2.110.0] - 2020-01-13
+
+Feature:
+
+- Add `buttonText` prop to `CrowdfundingCard`.
+
+Fix:
+
+- Fix overflow style on `ScrollableContainer`.
 
 ## [2.109.0] - 2020-01-12